% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.occurrence.R
\name{co.occurrence}
\alias{co.occurrence}
\title{Co-occurrence Analysis \code{cupcakeResults}}
\usage{
co.occurrence(input, pth, ageRange = c(0, 100), aggregate = TRUE,
  gender = "ALL", variation = c("", ""), nfactor = 10, scoreCutOff,
  fdrCutOff, oddsRatioCutOff, relativeRiskCutOff, phiCutOff, cores = 1,
  verbose = FALSE)
}
\arguments{
\item{input}{A \code{cupcakeData} object, obtained with the my.data function.}

\item{pth}{Determines the path where the required file with phenotype data is located.
This file is generated applying the \code{phenotypeSummary} function.}

\item{ageRange}{Determines what is the age range of interest for
performing the comorbidity analysis. By default it is set from 0 to 100 
years old.}

\item{aggregate}{By default TRUE. Change it to FALSE if you want to 
analyze the comorbidity taking into all the values of each phenotype.}

\item{gender}{Determine what is the gender of interest for 
performing the comorbidity analysis. By default \code{ALL}. Change it to the 
gender of interest for your comorbidity analysis.}

\item{variation}{Determine what is the variation of interest for 
performing the comorbidity analysis. By default \code{c("", "")}. Change it to the 
value of interest for your comorbidity analysis. For example, \code{c("CHD8", "yes")}}

\item{nfactor}{By default 10. Change it into other number if you consider there is any
categorical variable with more than nfactor values.}

\item{scoreCutOff}{The comorbidity score is a measure based on  the observed comorbidities
and the expected ones, based on the occurrence of each disease.}

\item{fdrCutOff}{A Fisher exact test for each pair of diseases is performed to assess 
the null hypothesis of independence between the two diseases. The Benjamini-Hochberg 
false discovery rate method (FDR) is applied to correct for multiple testing.}

\item{oddsRatioCutOff}{The odds ratio represents the increased chance that someone 
suffering disease X will have the comorbid disorder Y.}

\item{relativeRiskCutOff}{The relative risk refers to the fraction between the number of 
patients diagnosed with both diseases and random expectation based on disease 
prevalence.}

\item{phiCutOff}{The Pearsons correlation for binary variables (Phi) measures the 
robustness of the comorbidity association.}

\item{cores}{By default \code{1}. To run parallel computations on machines 
with multiple cores or CPUs, the cores argument can be changed.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
An object of class \code{cupcakeResults}
}
\description{
Given an object of type \code{cupcakeData}, a co-occurrence analysis is perform, 
for the subset of population under specific conditions of age, gender and gene status. 
It generates a \code{cupcakeResults} object.
}
\examples{
load(system.file("extdata", "genophenoExData.RData", package="Rcupcake"))
cooccurrenceExample <- co.occurrence( 
              input         = genophenoExData,
              pth           = system.file("extdata", package="Rcupcake"),
              aggregate     = TRUE, 
              ageRange      = c(0,16),
              gender        = "male", 
              )
}
