% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooc.heatmap.R
\name{cooc.heatmap}
\alias{cooc.heatmap}
\title{Plot the comorbidity analysis results in a heatmap.}
\usage{
cooc.heatmap(input, representedVariable = "patientsPhenoAB",
  variableCutOff = 0.05, coocPatients = 0, interactive = FALSE,
  lowColor = "#cde6ff", highColor = "red", verbose = FALSE)
}
\arguments{
\item{input}{A \code{cupcakeResults} object, obtained 
by applying the \code{comorbidityAnalysis} function}

\item{representedVariable}{By default \code{"patientsPhenoAB"} variable will be selected. Change
it to any of the other possible variables (\code{'score'},(\code{'fdr'},\code{'oddsRatio'}, 
\code{'phi'}, \code{'relativeRisk'}, \code{'PercentagePhenoAB'}).}

\item{variableCutOff}{By default \code{'0.05'}. The value of the argument can be changed 
to any other numeric variable, according to the range of the selected value.}

\item{coocPatients}{by default \code{'0'}.  The value of the argument can be changed
to any other numeric variable to show in the network only those comorbidities 
suffered by at least \code{coocPatients} of patients.}

\item{interactive}{Determines if the output heatmap is interactive or not. 
By default the \code{interactive} argument is set up as \code{FALSE}. The value 
of the argument can be changed to \code{TRUE}, as a result an interactive 
heatmap will be obtained.}

\item{lowColor}{Determines the heatmap color for the lowest value.
By default it is set to \code{"#cde6ff"}.}

\item{highColor}{Determines the heatmap color for the highest value.
By default it is set to \code{"red"}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
A heatmap
}
\description{
Given an object of class \code{cupcakeResults} obtained from a comorbidity analysis, 
a heatmap is obtained.
}
\examples{
load(system.file("extdata", "RcupcakeExResult.RData", package="genophenoR"))
htmp <- cooc.heatmap( input        = cupcakeResults, 
              representedVariable  = "patientsPhenoAB", 
              variableCutOff       = 1, 
              coocPatients         = 1
              )
htmp
}
