% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2cupcake.R
\name{dataframe2cupcake}
\alias{dataframe2cupcake}
\title{Transform a data.frame into a \code{cupcakeData} object}
\usage{
dataframe2cupcake(input, phenotypes, variants, age, gender,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{Determines the file with the complete path where the required 
input file is located. This input file must contain the "patient_id", two demographic variables, "Gender" and "Age",
and a list of phenotypes and variations}

\item{phenotypes}{Vector that contains the phenotype variables}

\item{variants}{Vector that contains the variants names}

\item{age}{Vector that contains the age variable}

\item{gender}{Vector that contains the gender variable}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{cupcakeData}
}
\description{
Given a tabulated file that contains the data in the 
correct format and generates a \code{cupcakeData} object.
}
\examples{
queryExample <- dataframe2cupcake( input      = paste0(system.file("extdata", package="Rcupcake"), 
                                                "/queryOutput.txt"),
                                   age        = "Age",
                                   gender     = "Gender",
                                   phenotypes = "Diabetes|Arthritis|LiverCancer|AnyCancer",
                                   variants   = "PCB153",
                                   verbose    = TRUE)
queryExample
}
