% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAnalysis-extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,}
\alias{cAnalysis-methods}
\title{Obtain the raw query from a \code{cAnalysis} object.}
\usage{
\S4method{extract}{cAnalysis}(object)
}
\arguments{
\item{object}{of class \code{cAnalysis}, \code{molecularcAnalysis}, 
\code{molecularComorbidity} or \code{comorbidity}}
}
\value{
A \code{data.frame} containing the raw result from the initial data 
(comorbidity)or comorbidity analysis (cAnalysis)
}
\description{
Obtain the raw query from a \code{cAnalysis} object.
}
\examples{
\dontrun{
#Being x an cAnalysis
qr <- extract(x) 
}
}
