% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{extract}
\alias{extract}
\title{Raw data from \code{genopheno} and \code{genophenoComor}.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{Object of class \code{genopheno} or \code{genophenoComor}}

\item{...}{NO USED}
}
\value{
A \code{data.frame} containing \code{data.frame}  with the results
from a query or a \code{data.frame} with the phenotype comorbidity result.
}
\description{
Obtain the raw data from a query  stored in a 
\code{genopheno} object or the raw data with all the phenotype
comorbidity measures of a \code{genophenoComor} object.
}
