% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genohenoZscore.R
\name{genohenoZscore}
\alias{genohenoZscore}
\title{A graphical summary of the phenotype values}
\usage{
genohenoZscore(input, cutOff = c(-2, 2), nfactor = 10, showTable = FALSE,
  outputFile = FALSE, path = "", verbose = FALSE)
}
\arguments{
\item{input}{Object of \code{genopheno} class.}

\item{cutOff}{Z-score cut-off to categorize the continuous variable. By default it is set 
to -2 and 2.}

\item{nfactor}{By default 10. Change it into other number if you consider there is any
categorical variable with more than nfactor values.}

\item{showTable}{By default FALSE. Change it into TRUE in order to visualize the table
with the ressults.}

\item{outputFile}{By default FALSE. Change it into TRUE in order to generate a file with the 
phenotypes and the values for each of them and an extra column to be filled by the user
for further analysis. The output file is called 'phenoSummary.txt'}

\item{path}{By default the working directory. Define the path where you want the file to 
be saved}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
A multiple graph containing a barplot with age distribution,  
a boxplot representing age distribution by gender and a pie chart representing 
gender distribution.
}
\description{
Given an object of class \code{genopheno}, a figure containing a barplot for each
phenotype is displayed. Each barplot shows the population percentage suffering each
type of the phenotypes according to the values it takes, and distinguishing between
those having or not a mutation. Furthermore, a data.frame with the numerical variables
is obtained.
}
\examples{
load(system.file("extdata", "genopheno.RData", package="genophenoR"))
genohenoZscore( input      = queryExample,
                verbose    = FALSE 
           )
}
