% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPhenoComorbidity.R
\name{genoPhenoComorbidity}
\alias{genoPhenoComorbidity}
\title{Comorbidity Analysis \code{genophenoComor}}
\usage{
genoPhenoComorbidity(input, pth, ageRange = c(0, 100), aggregate = TRUE,
  gender = "ALL", mutation = c("ALL", "ALL"), nfactor = 10, score, fdr,
  oddsRatio, rr, phi, cores = 1, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{A \code{genopheno} object, obtained with the queryPheno function.}

\item{pth}{Determines the path where the required input file with 
the yes/no phenotype data is located.}

\item{ageRange}{Determines what is the age range of interest for
performing the comorbidity analysis. By default it is set from 0 to 100 
years old.}

\item{aggregate}{By default TRUE. Change it to FALSE if you want to 
analyze the comorbidity taking into all the values of each phenotype.}

\item{gender}{Determine what is the gender of interest for 
performing the comorbidity analysis. By default \code{ALL}. Change it to the 
gender of interest for your comorbidity analysis.}

\item{mutation}{Determine what is the mutation value of interest for 
performing the comorbidity analysis. By default \code{c("ALL", "ALL")}. Change it to the 
value of interest for your comorbidity analysis. For example, \code{c("CHD8", "yes")}}

\item{nfactor}{By default 10. Change it into other number if you consider there is any
categorical variable with more than nfactor values.}

\item{score}{The comorbidity score is a measure based on  the observed comorbidities
and the expected ones, based on the occurrence of each disease.}

\item{fdr}{A Fisher exact test for each pair of diseases is performed to assess 
the null hypothesis of independence between the two diseases. The Benjamini-Hochberg 
false discovery rate method (FDR) is applied to correct for multiple testing.}

\item{oddsRatio}{The odds ratio represents the increased chance that someone 
suffering disease X will have the comorbid disorder Y.}

\item{rr}{The relative risk refers to the fraction between the number of 
patients diagnosed with both diseases and random expectation based on disease 
prevalence.}

\item{phi}{The Pearsons correlation for binary variables (Phi) measures the 
robustness of the comorbidity association.}

\item{cores}{By default \code{1}. To run parallel computations on machines 
with multiple cores or CPUs, the cores argument can be changed.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{cgpAnalysis}
}
\description{
Given an object of type \code{genopheno}, a comorbidity analysis is perform, 
for the subset of population under specific conditions of age and gender. It 
generates a \code{cgpAnalysis} object.
}
\examples{
load(system.file("extdata", "genopheno.RData", package="genophenoR"))
ex1 <- genoPhenoComorbidity( 
              input         = queryExample,
              pth           = system.file("extdata", package="genophenoR"),
              aggregate     = TRUE, 
              ageRange      = c(0,16),
              gender        = "MALE", 
              mutation      = c("CHD8", "ALL")
              )
}
