% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoPhenoHeatmap.R
\name{genoPhenoHeatmap}
\alias{genoPhenoHeatmap}
\title{Plot the comorbidity analysis results in a heatmap.}
\usage{
genoPhenoHeatmap(input, selectValue = "score", cutOff = 0.05, npairs = 0,
  interactive = FALSE, lowColor = "#cde6ff", highColor = "red",
  verbose = FALSE)
}
\arguments{
\item{input}{A \code{genophenoComor} object, obtained 
by applying the \code{comorbidityAnalysis} function}

\item{selectValue}{By default \code{"patientsPhenoAB"} variable will be selected. Change
it to any of the other possible variables (\code{'score'},(\code{'fdr'},\code{'odds ratio'}, 
\code{'phi'}, \code{'rr'}).}

\item{cutOff}{By default \code{'0.05'}. The value of the argument can be changed 
to any other numeric variable, according to the range of the selected value.}

\item{npairs}{by default \code{'0'}.  The value of the argument can be changed
to any other numeric variable to show in the network only those comorbidities 
suffered by at least \code{npairs} of patients.}

\item{interactive}{Determines if the output heatmap is interactive or not. 
By default the \code{interactive} argument is set up as \code{FALSE}. The value 
of the argument can be changed to \code{TRUE}, as a result an interactive 
heatmap will be obtained.}

\item{lowColor}{Determines the heatmap color for the lowest value.
By default it is set to \code{"#cde6ff"}.}

\item{highColor}{Determines the heatmap color for the highest value.
By default it is set to \code{"red"}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A heatmap
}
\description{
Given an object of class \code{genophenoComor} obtained from a comorbidity analysis, 
a heatmap is obtained.
}
\examples{
load(system.file("extdata", "genophenoComor.RData", package="genophenoR"))
htmp <- genoPhenoHeatmap( input = genophenoComor, 
              selectValue  = "patientsPhenoAB", 
              cutOff       = 1, 
              npairs       = 1
              )
htmp
}
