% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.children.R
\name{get.children}
\alias{get.children}
\title{Get children paths}
\usage{
get.children(fieldname, url, verbose = FALSE)
}
\arguments{
\item{fieldname}{The path in which the urser is interested.}

\item{url}{The url.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{apiKey}{The key to access to the data.}
}
\value{
A vector with all the fields that are under the path that has been given as input.
}
\description{
Given a url, a key and a path it returns all the children paths under the given one.
}
\examples{

nhanesPcbs <- get.children( 
                 fieldname   = "/nhanes/Demo/laboratory/laboratory/pcbs/", 
                 url         = "https://nhanes.hms.harvard.edu/"
              )
}
