% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.children.updated.R
\name{get.children.updated}
\alias{get.children.updated}
\title{Get children paths}
\usage{
get.children.updated(fieldname, url, verbose = FALSE)
}
\arguments{
\item{fieldname}{The path in which the user is interested.}

\item{url}{The url.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
A vector with all the fields that are under the path that has been given as input.
}
\description{
Given a url, a key and a path it returns all the children paths under the given one.
}
\examples{
nhanesPcbs <- get.children.updated( 
                 fieldname   = "/nhanes/Demo/laboratory/laboratory/pcbs/", 
                 url         = "https://nhanes.hms.harvard.edu/"
              )
}
