% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getchildren.R
\name{getchildren}
\alias{getchildren}
\title{Get children paths}
\usage{
getchildren(fieldname, verbose = FALSE)
}
\arguments{
\item{fieldname}{The path in which the urser is interested.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{url}{The url.}

\item{apiKey}{The key to access to the data.}
}
\value{
A vector with all the fields that are under the path that has been given as input.
}
\description{
Given a url, a key and a path it returns all the children paths under the given one.
}
\examples{

nhanesPcbs <- getchildren( 
                 fieldname   = "/nhanes/Demo/laboratory/laboratory/pcbs/"
              )
}
