% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irctquery.R
\name{irctquery}
\alias{irctquery}
\title{Query analysis to the API}
\usage{
irctquery(url, apiKey, jsonObject, verbose = FALSE)
}
\arguments{
\item{url}{The url.}

\item{apiKey}{The personal key to access to the data.}

\item{jsonObject}{A text file containing the JSON query body.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{data.frame}
}
\description{
Given an url, a key and a JSON object, it generates a \code{data.frame} object.
}
\examples{

query <- irctquery( 
              url         = "https://nhanes.hms.harvard.edu/", 
              apiKey      = "a77l42anvcgdtkfcvbl7hnp2v9"
              jsonObject  = system.file("extdata", "jsonQueryNhanes", package="genophenoR"))
              )
}
