% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.query.R
\name{my.query}
\alias{my.query}
\title{Query analysis to the API}
\usage{
my.query(myfields, myvector, url, verbose = FALSE,
  myfields.vector = NULL)
}
\arguments{
\item{myfields}{A vector with the fields of interest}

\item{myvector}{A vector with the paths of interest, generated applying the \code{getchildren}
function}

\item{url}{The url.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
A JSON query.
}
\description{
Given a vector with the fields of interest, and the vector generated with the paths obtained 
after applying the getchildren function, it returns a JSON query
}
\examples{
queryExample <- my.query( myfields = "AGE|PCB153",
}
