% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquery.R
\name{myquery}
\alias{myquery}
\title{Query analysis to the API}
\usage{
myquery(myfields, myvector, verbose = FALSE)
}
\arguments{
\item{myfields}{A vector with the fields of interest}

\item{myvector}{A vector with the paths of interest, generated applying the getchildren
function}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
A JSON query.
}
\description{
Given a vector with the fields of interest, and the paths vector generated applying
the getchildren function, it returns a JSON query
}
\examples{
nhanesPcbs <- getchildren( 
              url         = "https://nhanes.hms.harvard.edu/", 
              apiKey      = "a77l42anvcgdtkfcvbl7hnp2v9",
              fieldname   = "/nhanes/Demo/laboratory/laboratory/pcbs/"
              )
queryExample <- myquery( myfields = "AGE|PCB153", 
                         myvector = nhanesPcbs
              )
}
