% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{n.phenotype}
\alias{n.phenotype}
\title{Get the phenotypic variables from \code{cupcakeData}.}
\usage{
n.phenotype(object)
}
\arguments{
\item{object}{Object of class \code{cupcakeData}.}
}
\description{
Obtain the phenotypes in a \code{cupcakeData}.
}
\examples{
#Being qr an cupcakeData object
#n.phenotype(qr)
}
