% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/cupcake-n.phenotype.R
\docType{methods}
\name{n.phenotype}
\alias{n.phenotype}
\alias{n.phenotype,cupcakeData-method}
\title{Getter from \code{cupcakeData}.}
\usage{
n.phenotype(object)

\S4method{n.phenotype}{cupcakeData}(object)
}
\arguments{
\item{object}{Object of class \code{cupcakeData}.}
}
\value{
The number of unique phenotypes
}
\description{
Obtain the phenotypes in a \code{cupcakeData}.
}
\section{Methods (by class)}{
\itemize{
\item \code{cupcakeData}: get the distinct phenotypes
}}

\examples{
data(qr)
n.phenotype(qr)
}
