% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{n.variation}
\alias{n.variation}
\title{Get the alteration variables from \code{cupcakeData}.}
\usage{
n.variation(object)
}
\arguments{
\item{object}{Object of class \code{cupcakeData}.}
}
\description{
Obtain the alteration variables in a \code{cupcakeData}.
}
\examples{
data(qr)
n.variation(qr)
}
