% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{nmutation}
\alias{nmutation}
\title{Getter from \code{genopheno}.}
\usage{
nmutation(object)
}
\arguments{
\item{object}{Object of class \code{genopheno}.}
}
\description{
Obtain the mutations in a \code{genopheno}.
}
\examples{
data(qr)
nmutation(qr)
}
