% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/genopheno-nphenotype.R
\docType{methods}
\name{nphenotype}
\alias{nphenotype}
\alias{nphenotype,genopheno-method}
\title{Getter from \code{genopheno}.}
\usage{
nphenotype(object)

\S4method{nphenotype}{genopheno}(object)
}
\arguments{
\item{object}{Object of class \code{genopheno}.}
}
\value{
The number of unique phenotypes
}
\description{
Obtain the phenotypes in a \code{genopheno}.
}
\section{Methods (by class)}{
\itemize{
\item \code{genopheno}: get the distinc phenotypes
}}

\examples{
data(qr)
nphenotype(qr)
}
