% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.selection.R
\name{patient.selection}
\alias{patient.selection}
\title{Patients selection \code{genophenoPatientsSelection}}
\usage{
patient.selection(input, pth, ageRange = c(0, 100), phenotypeA, phenotypeB,
  aggregate = TRUE, gender = "ALL", variation = c("ALL", "ALL"),
  nfactor = 10, cores = 1, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{A \code{genopheno} object, obtained with the queryPheno function.}

\item{pth}{Determines the path where the required input file with 
the yes/no phenotype data is located.}

\item{ageRange}{Determines what is the age range of interest for
performing the comorbidity analysis. By default it is set from 0 to 100 
years old.}

\item{phenotypeA}{One of the phenotypes of interest and the value 
in which user is interested. For example, c("FacialExpression", "yes)}

\item{phenotypeB}{The second phenotype of interest and the value 
in which user is interested. For example, c("HandMovement", "yes)}

\item{aggregate}{By default TRUE. Change it to FALSE if you want to 
analyze the comorbidity taking into all the values of each phenotype.}

\item{gender}{Determine what is the gender of interest for 
performing the comorbidity analysis. By default \code{ALL}. Change it to the 
gender of interest for your comorbidity analysis.}

\item{variation}{Determine what is the variation value of interest for 
performing the comorbidity analysis. By default \code{c("ALL", "ALL")}. Change it to the 
value of interest for your comorbidity analysis. For example, \code{c("CHD8", "yes")}}

\item{nfactor}{By default 10. Change it into other number if you consider there is any
categorical variable with more than nfactor values.}

\item{cores}{By default \code{1}. To run parallel computations on machines 
with multiple cores or CPUs, the cores argument can be changed.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{cgpAnalysis}
}
\description{
Given an object of type \code{genopheno} and two phenotypes of
interest, the patients identifiers of those patients having
both phenotypes are selected.
}
\examples{
load(system.file("extdata", "cupcakeData.RData", package="Rcupcake"))
ex1 <- patient.selection( 
              input         = cupcakeData,
              pth           = system.file("extdata", package="Rcupcake"),
              phenotypeA    = c("Herpes", "yes"),
              phenotypeB    = c("HIV", "yes"),
              aggregate     = TRUE, 
              ageRange      = c(18,40),
              gender        = "male", 
              )
}
