% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.prevalence.R
\name{pheno.prevalence}
\alias{pheno.prevalence}
\title{Generates a table with the prevalence of each phenotype}
\usage{
pheno.prevalence(input, verbose = FALSE, warnings = FALSE)
}
\arguments{
\item{input}{A \code{cupcakeResults} object, obtained 
by applying the \code{co.occurrence} function}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A table
}
\description{
Given an object of class \code{cupcakeResults} obtained from the co-occurrence analysis, 
a prevalence table is obtained.
}
\examples{
load(system.file("extdata", "RcupcakeExResult.RData", package="Rcupcake"))
pheno.prevalence <- genoPhenoPrevalence( 
              input            = cupcakeResults
              )
}
