% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalencePlot.R
\name{prevalencePlot}
\alias{prevalencePlot}
\title{Plot the phenotype prevalence in a barplot (when there is not any mutation) and in a network 
if a mutation is defined}
\usage{
prevalencePlot(input, layout = "layout.circle",
  title = "Phenotype prevalence", mutation, interactive = FALSE,
  verbose = FALSE)
}
\arguments{
\item{input}{A \code{genophenoComor} object, obtained 
by applying the \code{comorbidityAnalysis} function}

\item{layout}{By default \code{'layout.fruchterman.reingold'}. It can be set 
to any other of the possible igraph layouts.}

\item{title}{Determines the title of the network figure. By default 
\code{'Comorbidity network'}.}

\item{mutation}{Determine what is the mutation value of interest for 
performing the comorbidity analysis.}

\item{interactive}{Determines if the output network is interactive or not. 
By default the \code{interactive} argument is set up as \code{FALSE}. The value 
of the argument can be changed to \code{TRUE}, as a result an interactive 
network will be obtained.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A barplot or a network
}
\description{
Given a \code{genophenoComor} object a barplot or a network is obtained.
}
\examples{
load(system.file("extdata", "genophenoComor.RData", package="genophenoR"))
ntwk <- prevalencePlot( 
              input            = genophenoComor 
              )
}
