% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGenoPheno.R
\name{queryGenoPheno}
\alias{queryGenoPheno}
\title{Query your data and generates a \code{genopheno}}
\usage{
queryGenoPheno(inputDataFile, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{inputDataFile}{Determines the file with the complete path where the required 
input file is located.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{phenotype}
}
\examples{
queryExample <- queryGenoPheno( inputDataFile = paste0(system.file("extdata", package="genophenoR"), 
                                                "/genophenoExData.txt"),
                                  verbose     = FALSE)
queryExample
}
