% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.query.R
\name{run.query}
\alias{run.query}
\title{Query analysis to the API}
\usage{
run.query(query, url, verbose = TRUE)
}
\arguments{
\item{query}{A JSON query, created with my.query function or contained in a
text file.}

\item{url}{The url.}

\item{verbose}{By default \code{TRUE}.}
}
\value{
An object of class \code{data.frame} with the query output.
}
\description{
Given an url and a JSON object, it generates a \code{data.frame} object with 
the output of the query.
}
\examples{

#query <- my.data( 
#              query  = system.file("extdata", "jsonQueryNhanes", package="Rcupcake"), 
#              url    = "https://nhanes.hms.harvard.edu/"
#              )
}
\author{
Alba Gutierrez, Gregoire Versmee
}
