% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.path.R
\name{search.path}
\alias{search.path}
\title{Search for a path}
\usage{
search.path(fieldname, url, verbose = FALSE)
}
\arguments{
\item{fieldname}{The path in which the user is interested.}

\item{url}{The url.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
The path for the folder of interest.
}
\description{
Given a url and a folder name it performs a breadth-first search through all the
available paths, and stops on the first occurence corresponding to the asked path
}
\examples{

pathExample <- search.path( 
                 fieldname   = "demographics", 
                 url         = "https://nhanes2.hms.harvard.edu/"
              )
}
