% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.request.R
\name{setToken}
\alias{setToken}
\title{Send a Request to the PIC-SURE API}
\usage{
setToken(t)
}
\arguments{
\item{url}{The domain name of the PIC-SURE API.}

\item{path}{The path of the url coming after the domain name}

\item{params}{The GET parameters to send}

\item{body}{The body to send with a POST request}
}
\value{
the returned value by Rcurl
}
\description{
Send a Request to the PIC-SURE API
}
