% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSession.R
\name{startSession}
\alias{startSession}
\title{Start the connection to the database}
\usage{
startSession(url, apiKey)
}
\arguments{
\item{url}{The url.}

\item{apiKey}{The key to access to the data.}
}
\value{
A message showing if the connection has been done or not.
}
\description{
Given an url and a key it start the connection to the database
}
\examples{

sessionEx <- startSession( 
              url         = "https://nhanes.hms.harvard.edu/", 
              apiKey      = "a77l42anvcgdtkfcvbl7hnp2v9"
              )
}
