% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPheno.R
\name{summaryPheno}
\alias{summaryPheno}
\title{A graphical summary of the database content.}
\usage{
summaryPheno(input, maleCode = "Male", femaleCode = "Female",
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{Object of \code{phenotype} class.}

\item{maleCode}{Characters(s) used to determine the male condition of a patient. 
Depending on the database it can be determined, for example, as \code{Male}, .
\code{MALE}, \code{M}, with digits as \code{0} or \code{1}.}

\item{femaleCode}{Characters(s) used to determine the female condition of a patient. 
Depending on the database it can be determined, for example, as \code{Female}, .
\code{FEMALE}, \code{F}, with digits as \code{0} or \code{1}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A multiple graph containing a barplot with age distribution,  
a boxplot representing age distribution by gender and a pie chart representing 
gender distribution.
}
\description{
Given an object of class \code{phenotype}, and the characters used to 
specify the gender, a figure containing 3 plots, one of the age distribution, 
another one of the age distribution and the third one with the relation 
between age and gender distribution.
}
\examples{
load(system.file("extdata", "phenogeno.RData", package="genophenoR"))
summaryPheno( input      = data2b2, 
           maleCode   = "MALE", 
           femaleCode = "FEMALE"
           )
}
