% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.score.R
\name{z.score}
\alias{z.score}
\title{Transform continuous in categorical variables and generates a new \code{cupcakeData} object.}
\usage{
z.score(input, zscoreCutOff = c(-2, 2), nfactor = 10, verbose = FALSE)
}
\arguments{
\item{input}{Object of \code{cupcakeData} class.}

\item{zscoreCutOff}{Z-score cut-off to categorize the continuous variable. By default it is set 
to -2 and 2.}

\item{nfactor}{By default 10. Change it into other number if you consider there is any
categorical variable with more than nfactor values.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
A \code{cupcakeData} class object with the continuous variable transformed into a categorical
variable, if possible.
}
\description{
Given an object of class \code{cupcakeData}, it transforms continuous into categorical variable 
applying Z-score. As a result a new \code{cupcakeData} object is generated. Note that if the number
of individuals is lower than 5000 a Saphiro test is done to test the normal distribution, otherwise
a Kolmogorov-Smirnov test is performed.
}
\examples{
load(system.file("extdata", "cupcakeDataExample.RData", package="Rcupcake"))
z.score( input      = cupcakeDataExample,
         verbose    = FALSE 
        )
}
