% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2cupcake.R
\name{dataframe2cupcake}
\alias{dataframe2cupcake}
\title{Transform a data.frame into a \code{cupcakeData} object}
\usage{
dataframe2cupcake(input, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{Determines the file with the complete path where the required 
input file is located. This input file must contain three columns: "patient_id" with
the patient identifier, "Gender" and "Age". Variation columns must start with "M." while
phenotype ones must start with a "P."}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{cupcakeData}
}
\description{
Given a tabulated file that contains the data in the 
correct format and generates a \code{cupcakeData} object.
}
\examples{
queryExample <- dataframe2cupcake( input = paste0(system.file("extdata", package="Rcupcake"), 
                                                "/queryOutput.txt"),
                                  verbose     = TRUE)
queryExample
}
