% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{extract}
\alias{extract}
\title{Obtain the raw query from a \code{genopheno} and \code{genophenoComor} object.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{of class \code{genophenoComor} or \code{genopheno} object}
}
\value{
A \code{data.frame} containing the raw result from the initial data 
(genopheno)or comorbidity analysis (genophenoComor)
}
\description{
Obtain the raw query from a \code{genopheno} and \code{genophenoComor} object.
}
\examples{
\dontrun{
#Being x an genophenoComor
qr <- extract(x) 
}
}
