% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genopheno-extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,}
\alias{genophenoComor-methods}
\title{Obtain the raw query from a \code{genophenoComor} object.}
\usage{
\S4method{extract}{genopheno}(object)
}
\arguments{
\item{object}{of class \code{genophenoComor} or \code{genopheno} object}
}
\value{
A \code{data.frame} containing the raw result from the initial data 
(genopheno)or comorbidity analysis (genophenoComor)
}
\description{
Obtain the raw query from a \code{genophenoComor} object.
}
\examples{
\dontrun{
#Being x an genophenoComor
qr <- extract(x) 
}
}
