% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irctquery.R
\name{irctquery}
\alias{irctquery}
\title{Query analysis to the API}
\usage{
irctquery(url, apiKey, query, outputPath = getwd(), verbose = FALSE)
}
\arguments{
\item{url}{The url.}

\item{apiKey}{The key to access to the data.}

\item{query}{A text file containing the JSON query body.}

\item{outputPath}{Path where the output file will be saved. By default it will be 
saved in your working directory}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
An object of class \code{data.frame} with the query output.
}
\description{
Given an url, a key and a JSON object, it generates a \code{data.frame} object with 
the output of the query.
}
\examples{

query <- irctquery( 
              url         = "https://nhanes.hms.harvard.edu/", 
              apiKey      = "a77l42anvcgdtkfcvbl7hnp2v9", 
              query  = system.file("extdata", "jsonQueryNhanes", package="genophenoR"))
              )
}
