% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.query.JHS.R
\name{my.query.JHS}
\alias{my.query.JHS}
\title{!!!!!!! Fix an issue with The Jackson Heart Study Cohort on Topmed-dev
Query analysis to the API}
\usage{
my.query.JHS(myfields, myvector, url, verbose = FALSE)
}
\arguments{
\item{myfields}{A vector with the fields of interest}

\item{myvector}{A vector with the paths of interest, generated applying the \code{getchildren}
function}

\item{url}{The url.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an on-time log from the function.}
}
\value{
A JSON query.
}
\description{
Given a vector with the fields of interest, and the vector generated with the paths obtained 
after applying the getchildren function, it returns a JSON query
}
\author{
Gregoire Versmee, Gabor Korodi
}
