% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.selection.R
\name{patient.selection}
\alias{patient.selection}
\title{Patients selection}
\usage{
patient.selection(input, pth, ageRange = c(0, 100), phenotypeA, phenotypeB,
  aggregate = TRUE, gender = "ALL", variation = c("ALL", "ALL"),
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{A \code{cupcakeData} object, obtained with the \code{my.data} function.}

\item{pth}{Determines the path where the required input file with 
the phenotypic data is located.}

\item{ageRange}{Determines what is the age range of interest for
performing the co-occurrence analysis. By default it is set from 0 to 100 
years old.}

\item{phenotypeA}{One of the phenotypes of interest and the value 
in which user is interested. For example, c("Arthritis", "Yes")}

\item{phenotypeB}{The second phenotype of interest and the value 
in which user is interested. For example, c("Diabetes", "Yes")}

\item{aggregate}{By default TRUE. Change it to FALSE if you want to 
analyze the co-occurrence taking into all the values of each phenotype.}

\item{gender}{Determine what is the gender of interest for 
performing the co-occurrence analysis. By default \code{ALL}. Change it to the 
gender of interest for your co-occurrence analysis.}

\item{variation}{Determine what is the variation value of interest for 
performing the co-occurrence analysis. By default \code{c("ALL", "ALL")}. Change it to the 
value of interest for your co-occurrence analysis. For example, \code{c("CHD8", "yes")}}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{cgpAnalysis}
}
\description{
Given an object of type \code{cupcakeData} and two phenotypes of
interest, the patients having both phenotypes are selected.
}
\examples{
load(system.file("extdata", "RcupcakeDataExData.RData", package="Rcupcake"))
ex1 <- patient.selection( 
              input         = RcupcakeExData,
              pth           = system.file("extdata", package="Rcupcake"),
              phenotypeA    = c("Arthritis", "Yes"),
              phenotypeB    = c("Diabetes", "Yes"),
              aggregate     = TRUE, 
              ageRange      = c(0,100),
              gender        = "male", 
              )
}
