% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeSummary.R
\name{phenotypeSummary}
\alias{phenotypeSummary}
\title{Describes the phenotypic characteristics for the whole study population. If gene is selected, 
the summary will contain the results regardin the selected gene.}
\usage{
phenotypeSummary(input, mutation = FALSE, nfactor = 10, showTable = TRUE,
  showFigures = FALSE, path = getwd(), verbose = FALSE)
}
\arguments{
\item{input}{Object of \code{genopheno} class.}

\item{mutation}{Determines the mutation of interest for which you want to analyze the 
phenotype values. By default FALSE.}

\item{nfactor}{By default 10. Change it into other number if you consider there is any
categorical variable with more than \code{nfactor} values.}

\item{showTable}{By default TRUE. Change it into FALSE in order to not visualize the table
with the results.}

\item{showFigures}{By default FALSE. Change it into TRUE in order to visualize the table
with the results.}

\item{path}{By default the working directory. Define the path where you want the file to 
be saved}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}
}
\value{
A file .
}
\description{
Given an object of class \code{genopheno}, a file with the different values for each phenotype, 
and the prevalence of each one in general population and according to the gene status, if present,
is generated. A figure containing a barplot or boxplot for each phenotype is displayed. 
Each barplot shows the population percentage suffering each type of the phenotypes according to the values it takes, 
and distinguishing between those having or not a mutation, if present. 
Furthermore, a data.frame with the numerical values is obtained.
}
\examples{
load(system.file("extdata", "genophenoExData.RData", package="genophenoR"))
phenotypeSummary( input       = genophenoExData, 
                  showTable   = TRUE, 
                  showFigures = TRUE, 
                 verbose      = FALSE 
           )
}
