% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.plot.R
\name{prevalence.plot}
\alias{prevalence.plot}
\title{Plot the phenotype prevalence in a barplot (when there is not any variation) and in a network 
if a variation is defined}
\usage{
prevalence.plot(input, layout = "layout.circle", variation, verbose = FALSE)
}
\arguments{
\item{input}{A \code{cupcakeResults} object, obtained 
by applying the \code{co.occurrence} function}

\item{layout}{By default \code{'layout.circle'}. It can be set 
to any other of the possible igraph layouts.}

\item{variation}{Determine what is the variation value of interest for 
performing the co-occurrence analysis.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get an
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A barplot or a network
}
\description{
Given a \code{cupcakeResults} object a barplot or a network is obtained.
}
\examples{
load(system.file("extdata", "RcupcakeExResult.RData", package="Rcupcake"))
ntwk <- prevalence.plot( 
              input            = cupcakeResults 
              )
}
