% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGenoPheno.R
\name{queryGenoPheno}
\alias{queryGenoPheno}
\title{Query your data and generates a \code{phenoytpe}}
\usage{
queryGenoPheno(databasePth, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{databasePth}{Determines the path where the required input file 
islocated.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{phenotype}
}
\examples{
ex1 <- queryPheno( databasePth = system.file("extdata", package="phenotypeR"),
                   verbose     = FALSE)
}
